requirements.txt
kafka-python-ng
faker
